Feat: 댓글에 이모티콘 반응 달기 기능 구현 (#62)
* style: 스켈레톤 색상 수정

* fix: modal provider 상태관리 수정

- 기존 provider 상태로는 한번에 모든 모달이 열리는 이슈가있어서 수정함

* feat: icon 추가

* feat: ReactQueryDevtools 설치 및 세팅

* feat: Modal index 추가

* feat: Button에 새 icon size props 추가

* fix: 누락된 key 추가

* chore: 필요없는 주석 삭제

* feat: 댓글 삭제, 수정 기능 구현

* chore: 필요없는 주석 삭제

* chore: 누락된 커밋

* fix: 삭제 모달 modal 리팩토링한거에 맞춰서 수정

* fix: 누락된 'use client' 추가

* refactor: 이모티콘 피드백 페이지를 server component로 수정

* feat: 이모티콘 이미지 6개(첫줄) priority 부여

* fix: hydration 에러 수정

* feat: 이모티콘 이미지들 첫줄의 fetchPriorty,loading 값 차별화

* chore: kakao -> kakaotalk으로 수정

* feat: 댓글 반응 api 구현

* feat: emoticon set detail api에 코멘트 갯수 추가

* refactor: 코멘트 쿼리 키 가독성 좋게 수정

* feat: comment api에 댓글 반응 값 추가

* fix: kakao->kakaotalk 수정 반영

* feat: 댓글 섹션을 클라이언트 컴포넌트로 수정

* feat: 댓글에 이모지 반응 달기,지우기 기능 추가

* refactor: render 함수 외부로 분리

* chore: react-scan, zustand 라이브러리 설치

* feat: react scan 추가

* chore: emoticon-comment-section 파일 위치 이동

* feat: comment ui 상태를 provider로 관리

* refactor: emoticon ui section 상태를 useSyncExternalStore로 관리

* feat: 수정 중일 때 다시 showmore 버튼 누를경우 수정 모드 풀리도록 추가

* refactor: Comment 컴포넌트 분리

* feat: 댓글 리액션 낙관적 업데이트로 수정

* feat: popover들 밖부분들 누르면 닫히게 추가

+++ src/entity/comment/api/comment-api.ts
@@ -15,11 +15,13 @@ export async function getComments(
   try {
     const supabase = await createServerSupabaseClient();
 
-    let query = supabase
-      .from('comments')
-      .select(
-        '*, profile:profiles!comments_user_id_fkey(id, nickname, avatar_url)',
-      );
+    let query = supabase.from('comments').select(
+      `
+        *,
+        profile:profiles!comments_user_id_fkey(*),
+        comment_reactions!comment_reactions_comment_id_fkey(*)
+        `,
+    );
 
     if (request?.set_id) {
       query = query.eq('set_id', request?.set_id);
@@ -30,7 +32,11 @@ export async function getComments(
     }
 
     if (request?.parent_comment_id !== undefined) {
-      query = query.eq('parent_comment_id', request.parent_comment_id || '');
+      if (request.parent_comment_id === null) {
+        query = query.is('parent_comment_id', null);
+      } else {
+        query = query.eq('parent_comment_id', request.parent_comment_id);
+      }
     }
 
     if (request.sortBy) {
@@ -57,10 +63,14 @@ export async function getComments(
           countQuery = countQuery.eq('user_id', request.user_id);
         }
         if (request.parent_comment_id !== undefined) {
-          countQuery = countQuery.eq(
-            'parent_comment_id',
-            request.parent_comment_id || '',
-          );
+          if (request.parent_comment_id === null) {
+            countQuery = countQuery.is('parent_comment_id', null);
+          } else {
+            countQuery = countQuery.eq(
+              'parent_comment_id',
+              request.parent_comment_id,
+            );
+          }
         }
 
         return countQuery;
@@ -89,12 +99,28 @@ export async function getComments(
       };
     }
 
-    const validComments = (dataResult.data || [])
-      .filter((item) => item.profile !== null)
-      .map((item) => ({
+    const validComments = dataResult.data;
+
+    const formatedComments = validComments?.map((item) => {
+      const reactionCounts = item.comment_reactions.reduce(
+        (acc, reaction) => {
+          acc[reaction.emoji] = (acc[reaction.emoji] || 0) + 1;
+          return acc;
+        },
+        {} as Record<string, number>,
+      );
+
+      return {
         ...item,
-        profile: item.profile!,
-      }));
+        reaction_summary: Object.entries(reactionCounts).map(
+          ([emoji, count]) => ({
+            emoji,
+            count,
+          }),
+        ),
+        reactions: item.comment_reactions,
+      };
+    });
 
     const total = countResult.count || 0;
     const currentPage = Math.floor(offset / limit) + 1;
@@ -104,7 +130,10 @@ export async function getComments(
     return {
       success: true,
       data: {
-        data: validComments,
+        data: formatedComments?.map((item) => ({
+          ...item,
+          profile: item.profile!,
+        })),
         total,
         currentPage,
         totalPages,
@@ -137,8 +166,8 @@ export async function createComment(
 
   const commentData = {
     content: request.comment.content,
-    set_id: request.comment.set_id || '',
-    parent_comment_id: request.comment.parent_comment_id,
+    set_id: request.comment.set_id || null,
+    parent_comment_id: request.comment.parent_comment_id || null,
     image_id: request.comment.image_id,
     images: request.comment.images,
     user_id: user.id,
@@ -165,7 +194,6 @@ export async function updateComment(
   params: UpdateCommentParams,
 ): Promise<void> {
   const supabase = await createServerSupabaseClient();
-
   const { error } = await supabase
     .from('comments')
     .update({
@@ -173,7 +201,7 @@ export async function updateComment(
       images: params.images,
       updated_at: params.updated_at || new Date().toISOString(),
     })
-    .eq('id', params.id);
+    .eq('id', params.commentId);
 
   if (error) {
     throw error;
@@ -185,10 +213,14 @@ export async function deleteComment(
 ): Promise<void> {
   const supabase = await createServerSupabaseClient();
 
+  if (!params.commentId) {
+    throw new Error('commentId가 없습니다.');
+  }
+
   const { error } = await supabase
     .from('comments')
     .delete()
-    .eq('id', params.id);
+    .eq('id', params.commentId);
 
   if (error) {
     throw error;
